pyproject.toml
[project]
name = "rationai-ciao"
version = "0.1.0"
description = "CIAO: Contextual Importance Assessment via Obfuscation - An XAI method for identifying influential image regions"
authors = [{ name = "David Halmazňa", email = "david.halmazna@mail.muni.cz" }]
requires-python = ">=3.11,<3.14"
readme = "README.md"
license = { file = "LICENSE" }
dependencies = [
    # Core ML/DL frameworks
    "torch>=2.0.0",
    "torchvision>=0.15.0",

    # Configuration and experiment tracking
    "hydra-core>=1.3.0",
    "mlflow>=3.0",
    "omegaconf>=2.3.0",

    # XAI and visualization
    "matplotlib>=3.5.0",
    "plotly>=5.0.0",
    "ipywidgets>=7.0.0",
    "captum>=0.7.0",

    # Image processing and segmentation
    "scikit-image>=0.19.0",
    "pillow>=9.0.0",

    # Scientific computing
    "numpy>=1.21.0",

    # Others
    "tqdm>=4.0.0",
]

[project.scripts]
ciao = "ciao.__main__:main"
ciao-ep = "ciao.baselines.extremal.__main__:main"
ciao-lime = "ciao.baselines.lime.__main__:main"

[dependency-groups]
dev = ["mypy", "ruff"]
test = ["pytest", "pytest-cov"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["ciao"]

[tool.uv]
package = true
